Delete dead code, add test
