+++ server/package.json
@@ -28,6 +28,7 @@
     },
     "dependencies": {
         "@types/cors": "^2.8.6",
+        "@types/mocha": "^5.2.7",
         "async": "^3.1.0",
         "bcrypt-nodejs": "^0.0.3",
         "bluebird": "^3.5.5",
@@ -71,7 +72,7 @@
         "@types/express": "^4.17.1",
         "@types/express-flash": "0.0.1",
         "@types/express-session": "^1.15.14",
-        "@types/jest": "^24.0.18",
+        "@types/jest": "^24.0.24",
         "@types/jquery": "^3.3.31",
         "@types/lodash": "^4.14.141",
         "@types/lusca": "^1.6.1",
