+++ server/src/app.ts
@@ -15,18 +15,11 @@ import { MONGODB_URI, SESSION_SECRET } from "./util/secrets";
 const MongoStore = mongo(session);
 
 // Controllers (route handlers)
-import * as homeController from "./controllers/home";
 import * as userController from "./controllers/user";
-import * as apiController from "./controllers/api";
-import * as contactController from "./controllers/contact";
 import * as searchController from "./controllers/search";
 
-import * as signUp from "./controllers/signUp";
-
-
-// API keys and Passport configuration
+// Passport configuration
 import * as passportConfig from "./config/passport";
-import { nextTick } from "async";
 
 // Create Express server
 const app = express();
@@ -39,7 +32,6 @@ mongoose.connect(mongoUrl, { useNewUrlParser: true, useCreateIndex: true, useUni
     () => { /** ready to use. The `mongoose.connect()` promise resolves to undefined. */ },
 ).catch(err => {
     console.log("MongoDB connection error. Please make sure MongoDB is running. " + err);
-    // process.exit();
 });
 
 // Express configuration
@@ -69,8 +61,7 @@ const corsOptions = {
 };
 
 app.use(cors(corsOptions));
-// app.use(lusca.xframe("SAMEORIGIN"));
-// app.use(lusca.xssProtection(true));
+
 app.use((req, res, next) => {
     res.locals.user = req.user;
     next();
@@ -92,31 +83,21 @@ app.use((req, res, next) => {
 app.use(
     express.static(path.join(__dirname, "public"), { maxAge: 31557600000 })
 );
-// app.use((req, res, next) => {
-//     res.header("Access-Control-Allow-Origin", "http://localhost:3000/"); // update to match the domain you will make the request from
-//     res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-//     next();
-// });
+
 
 /**
  * Primary app routes.
  */
-// app.options("*");
-//app.get("/", homeController.index);
-app.get("/login", userController.getLogin);
+
 app.post("/login", userController.postLogin);
 app.get("/logout", userController.logout);
-app.get("/signup", userController.getSignup);
-app.post("/signup", signUp.postSignup);
+app.post("/signup", userController.postSignup);
 app.post("/searchUser", searchController.postSearchUser);
 app.post("/searchDoctor", searchController.postSearchDoctor);
-app.get("/contact", contactController.getContact);
-app.post("/contact", contactController.postContact);
 app.get("/getAllUser", passportConfig.isAuthenticated, searchController.getAllUser);
 app.get("/getAllDoctor", passportConfig.isAuthenticated, searchController.getAllDoctor);
 app.get("/account", passportConfig.isAuthenticated, userController.getAccount);
 app.post("/account/profile", passportConfig.isAuthenticated, userController.postUpdateProfile);
-app.post("/account/password", passportConfig.isAuthenticated, userController.postUpdatePassword);
 app.post("/account/delete", passportConfig.isAuthenticated, userController.postDeleteAccount);
 app.post("/account/add-record", passportConfig.isAuthenticated, userController.postAddRecord);
 app.get("/account/get-recordlist", passportConfig.isAuthenticated, userController.getRecordList);
@@ -126,20 +107,5 @@ app.post("/view-authorized-users", passportConfig.isAuthenticated, userControlle
 app.post("/view-user-records", passportConfig.isAuthenticated, userController.ViewuserRecord);
 app.get("/view-authorizing-users", passportConfig.isAuthenticated, userController.ViewAuthedUser);
 app.post("/remove-auth", passportConfig.isAuthenticated, userController.RemoveAuth);
-
-//app.use(cors(corsOptions));
-
-/**
- * API examples routes.
- */
-//app.get("/api", apiController.getApi);
-
-/**
- * OAuth authentication routes. (Sign in)
- */
-// app.get("/auth/facebook", passport.authenticate("facebook", { scope: ["email", "public_profile"] }));
-// app.get("/auth/facebook/callback", passport.authenticate("facebook", { failureRedirect: "/login" }), (req, res) => {
-//     res.redirect(req.session.returnTo || "/");
-// });
 
 export default app;
