--- server/src/controllers/home.ts
@@ -1,17 +1,0 @@
-import { Request, Response } from "express";
-
-/**
- * GET /
- * Home page.
- */
-export const index = (req: Request, res: Response) => {
-    if (req.user) {
-        console.log("HOME WITH USER");
-        return res.send({user: req.user});
-    }
-    console.log("HOME WITHOUT USER");
-    return res.send();
-    // res.render("home", {
-    //     title: "Home"
-    // });
-};
