--- server/src/controllers/signUp.ts
@@ -1,45 +1,0 @@
-import async from "async";
-import crypto from "crypto";
-import nodemailer from "nodemailer";
-import passport from "passport";
-import { AuthToken } from "../models/Admin";
-import { User, UserDocument } from "../models/User";
-import { Request, Response, NextFunction } from "express";
-import { IVerifyOptions } from "passport-local";
-import { WriteError } from "mongodb";
-import { check, sanitize, validationResult } from "express-validator";
-import "../config/passport";
-
-// Controllers (route handlers)
-import * as userFactory from "./user_factory";
-
-/**
- * POST /signup
- * Create a new local account.
- */
-export const postSignup = (req: Request, res: Response, next: NextFunction) => {
-
-    /* Global steps */
-
-
-    check("email", "Email is not valid").isEmail();
-    check("password", "Password must be at least 4 characters long").isLength({ min: 4 });
-    check("confirmPassword", "Passwords do not match").equals(req.body.password);
-    // eslint-disable-next-line @typescript-eslint/camelcase
-    sanitize("email").normalizeEmail({ gmail_remove_dots: false });
-
-    /** 
-     * Add an error handler for usertype?
-     */ 
-
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-        console.log(errors.array());
-        return res.status(400).send(errors);
-    }
-
-    /* User factory */
-
-    userFactory.createUser(req, res, next);
-};
