+++ server/src/models/Personnel.ts
@@ -1,5 +1,4 @@
 import bcrypt from "bcrypt-nodejs";
-import crypto from "crypto";
 import mongoose, { Document } from "mongoose";
 //import { string } from "prop-types";
