--- server/test/api.test.ts
@@ -1,9 +1,0 @@
-import request from "supertest";
-import app from "../src/app";
-
-describe("GET /api", () => {
-    it("should return 200 OK", () => {
-        return request(app).get("/api")
-            .expect(200);
-    });
-});
