--- server/test/app.test.ts
@@ -1,9 +1,0 @@
-import request from "supertest";
-import app from "../src/app";
-
-describe("GET /random-url", () => {
-    it("should return 404", (done) => {
-        request(app).get("/reset")
-            .expect(404, done);
-    });
-});
