--- server/test/home.test.ts
@@ -1,9 +1,0 @@
-import request from "supertest";
-import app from "../src/app";
-
-describe("GET /", () => {
-    it("should return 200 OK", (done) => {
-        request(app).get("/")
-            .expect(200, done);
-    });
-});
