--- server/views/account/forgot.pug
@@ -1,16 +1,0 @@
-extends ../layout
-
-block content
-  .col-sm-8.offset-sm-2
-    form(method='POST')
-      legend Forgot Password
-      hr
-      input(type='hidden', name='_csrf', value=_csrf)
-      .form-group
-        p Enter your email address below and we will send you password reset instructions.
-        label.col-form-label.font-weight-bold(for='email') Email
-        input.form-control(type='email', name='email', id='email', placeholder='Email', autofocus, required)
-      .form-group
-        button.btn.btn-primary(type='submit')
-          i.fa.fa-key
-          | Reset Password
