--- server/views/api/index.pug
@@ -1,13 +1,0 @@
-extends ../layout
-
-block content
-  h2 API Examples
-  hr
-
-  .row
-    .col-sm-4
-      a(href='/api/facebook', style='color: #fff')
-        .card(style='background-color: #3b5998')
-          .card-body
-            img(src='http://i.imgur.com/jiztYCH.png', height=40, alt='Facebook')
-            |  Facebook
