--- server/views/contact.pug
@@ -1,26 +1,0 @@
-extends layout
-
-block content
-  .page-header
-    h3 Contact Us
-    hr
-
-  form.form-horizontal(method='POST')
-    input(type='hidden', name='_csrf', value=_csrf)
-    .form-group.row.justify-content-md-center
-      label.col-sm-2.col-form-label.text-right.font-weight-bold(for='name') Name
-      .col-sm-8
-        input.form-control(type='text', name='name', id='name', autofocus=true)
-    .form-group.row.justify-content-md-center
-      label.col-sm-2.col-form-label.text-right.font-weight-bold(for='email') Email
-      .col-sm-8
-        input.form-control(type='text', name='email', id='email')
-    .form-group.row.justify-content-md-center
-      label.col-sm-2.col-form-label.text-right.font-weight-bold(for='message') Body
-      .col-sm-8
-        textarea.form-control(name='message', id='message', rows='7')
-    .form-group.row.justify-content-md-center
-      .offset-sm-2.col-sm-8
-        button.btn.btn-primary(type='submit')
-          i.fa.fa-envelope
-          | Send
