--- server/views/home.pug
@@ -1,18 +1,0 @@
-extends layout
-
-block content
-  p.title-1 WellBee
-  p.title-2 Electrion Health Record System
-  p.titlep A personal health data solution for you and those who you love.
-  hr
-  p
-  p
-  p
-  .row
-    .col-sm-6
-      h2 More Than An Archive
-      p WellBee helps users maintain a healthy lifestyle in a more efficient way by providing a quick analysis of their health, recommendations, and a powerful solution to doctor-patient communication.
-      p &nbsp
-      p
-        a.btn.btn-secondary(href='#', role='button') View more »
-
