--- server/views/layout.pug
@@ -1,32 +1,0 @@
-doctype html
-html(lang='')
-  head
-    meta(charset='utf-8')
-    meta(http-equiv='X-UA-Compatible', content='IE=edge')
-    meta(name='viewport', content='width=device-width, initial-scale=1.0')
-    title #{title} - WellBee
-    meta(name='description', content='')
-    meta(name='theme-color' content='#4DA5F4')
-    meta(name='csrf-token', content=_csrf)
-    link(rel='shortcut icon', href='/images/favicon.png')
-    link(rel='stylesheet', href='/css/main.css')
-
-  body
-    include partials/header
-
-    .container.pl-sm-0.pr-sm-0.pl-lg-3.pr-lg-3
-      include partials/flash
-      block content
-
-    include partials/footer
-
-    script(src='/js/lib/jquery-3.1.1.min.js')
-    script(src='/js/lib/bootstrap.min.js')
-    script(async defer src="https://buttons.github.io/buttons.js")
-    script(src='/js/main.js')
-
-    //- Google Analytics: change UA-38177297-2 to be your site's ID
-    script.
-      window.ga=function(){ga.q.push(arguments)};ga.q=[];ga.l=+new Date;
-      ga('create','UA-38177297-2','auto');ga('send','pageview')
-    script(src='https://www.google-analytics.com/analytics.js' async defer)
