--- server/views/partials/footer.pug
@@ -1,4 +1,0 @@
-footer
-  .container.text-center
-    p.pull-left © 2019 WellBee, Inc. All Rights Reserved
-    //a.github-button.float-right(href="https://github.com/CaseyPYZ/WellBee-nyush-SE-f2019/", data-icon="octicon-star", data-show-count="true", aria-label="Star microsoft/typescript-node-starter on GitHub") Star
